# api/swagger.yaml
...
swagger: '2.0'
info:
  description: 'SRS and workflow: https://docs.google.com/document/d/17aiIiosJApvnAliPajrcKnusvc9YdgLgiccv_TbMRGM

'
  version: 1.0.0
  title: Monero Wallet service API
basePath: "/"
schemes:
- http
- https
paths:
...
  "/api/v1/extra_multisig_info":
    post:
      summary: extra multisig info
      description: Second round of keys exchange
      tags:
      - Key exchange
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: X-Session-Id
        in: header
        required: true
        type: string
      - name: X-Nonce
        in: header
        required: true
        type: integer
        minimum: 0
      - name: X-Signature
        in: header
        required: true
        type: string
      - name: body
        in: body
        required: true
        schema:
          type: object
          required:
          - extra_multisig_info
          properties:
            extra_multisig_info:
              type: string
              description: includes public spend key + array of multisig keys + signature
      responses:
        '204':
          description: Multisig accepted
        '400':
          description: Status 400
          schema:
            required:
            - error
            type: object
            properties:
              error:
                type: string
        '403':
          description: Status 403
...
  "/api/v1/create_wallet":
    post:
      summary: create wallet
      tags:
      - Wallet
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: X-Session-Id
        in: header
        required: true
        type: string
      - name: X-Nonce
        in: header
        required: true
        type: integer
        minimum: 0
      - name: X-Signature
        in: header
        required: true
        type: string
      - name: body
        in: body
        required: true
        schema:
          type: object
          required:
          - multisig_info
          - name
          - participants
          - signers
          properties:
            participants:
              type: integer
              minimum: 2
            signers:
              type: integer
              minimum: 1
            name:
              type: string
            multisig_info:
              type: string
              description: view secret key + public spend key + signature
      responses:
        '200':
          description: Status 200
          schema:
            type: object
            properties:
              invite_code:
                type: string
                description: invite code to pass to other participants
        '400':
          description: Status 400
          schema:
            required:
            - error
            type: object
            properties:
              error:
                type: string
        '403':
          description: Status 403
  "/api/v1/wallet_invite":
    get:
      summary: get wallet invite code
      tags:
      - Wallet
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: X-Session-Id
        in: header
        required: true
        type: string
      - name: X-Nonce
        in: header
        required: true
        type: integer
        minimum: 0
      - name: X-Signature
        in: header
        required: true
        type: string
      responses:
        '200':
          description: Status 200
          schema:
            type: object
            properties:
              invite_code:
                type: string
                description: invite code to pass to other participants
        '400':
          description: Status 400
          schema:
            required:
            - error
            type: object
            properties:
              error:
                type: string
        '403':
          description: Status 403
  "/api/v1/join_wallet":
    post:
      summary: join wallet
      tags:
      - Wallet
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: X-Session-Id
        in: header
        required: true
        type: string
      - name: X-Nonce
        in: header
        required: true
        type: integer
        minimum: 0
      - name: X-Signature
        in: header
        required: true
        type: string
      - name: body
        in: body
        required: true
        schema:
          type: object
          required:
          - invite_code
          - multisig_info
          properties:
            invite_code:
              type: string
            multisig_info:
              type: string
              description: view secret key + public spend key + signature
      responses:
        '204':
          description: Joined successfully
        '400':
          description: Status 400
          schema:
            required:
            - error
            type: object
            properties:
              error:
                type: string
        '403':
          description: Status 403
...
  "/api/v2/create_wallet":
    post:
      summary: create wallet
      tags:
      - V2
        - Wallet
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: X-Session-Id
        in: header
        required: true
        type: string
      - name: X-Nonce
        in: header
        required: true
        type: integer
        minimum: 0
      - name: X-Signature
        in: header
        required: true
        type: string
      - name: body
        in: body
        required: true
        schema:
          type: object
          required:
          - participants
          - signers
          - supported_protocols
          properties:
            participants:
              type: integer
              minimum: 2
            signers:
              type: integer
              minimum: 1
            supported_protocols:
              type: array
              items:
                type: string
      responses:
        '200':
          description: Status 200
          schema:
            type: object
            properties:
              invite_code:
                type: string
                description: invite code to pass to other participants
        '400':
          description: Status 400
          schema:
            required:
            - error
            type: object
            properties:
              error:
                type: string
        '403':
          description: Status 403
...